@@ -19,13 +19,12 @@ import android.app.AlertDialog;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
-import android.content.SharedPreferences;
 import android.graphics.Color;
 import android.graphics.drawable.Drawable;
-import android.net.Uri;
 import android.os.Bundle;
 import android.preference.PreferenceManager;
 import android.support.annotation.AttrRes;
+import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.design.widget.AppBarLayout;
 import android.support.design.widget.CollapsingToolbarLayout;
@@ -41,31 +40,33 @@ import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
+import android.view.ViewGroup;
+import android.widget.Button;
 import android.widget.ImageView;
 import android.widget.TextView;
 
+import com.keylesspalace.tusky.db.AccountEntity;
 import com.keylesspalace.tusky.entity.Account;
 import com.keylesspalace.tusky.entity.Relationship;
+import com.keylesspalace.tusky.interfaces.ActionButtonActivity;
 import com.keylesspalace.tusky.interfaces.LinkListener;
 import com.keylesspalace.tusky.pager.AccountPagerAdapter;
 import com.keylesspalace.tusky.receiver.TimelineReceiver;
-import com.keylesspalace.tusky.util.LinkHelper;
 import com.keylesspalace.tusky.util.Assert;
+import com.keylesspalace.tusky.util.LinkHelper;
 import com.keylesspalace.tusky.util.ThemeUtils;
 import com.pkmmte.view.CircularImageView;
 import com.squareup.picasso.Picasso;
 
-import java.text.NumberFormat;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 
-import butterknife.BindView;
-import butterknife.ButterKnife;
 import retrofit2.Call;
 import retrofit2.Callback;
 import retrofit2.Response;
 
-public class AccountActivity extends BaseActivity {
+public final class AccountActivity extends BaseActivity implements ActionButtonActivity {
     private static final String TAG = "AccountActivity"; // logging tag
 
     private enum FollowState {
@@ -81,18 +82,37 @@ public class AccountActivity extends BaseActivity {
     private boolean isSelf;
     private Account loadedAccount;
 
-    @BindView(R.id.account_avatar) CircularImageView avatar;
-    @BindView(R.id.account_header) ImageView header;
-    @BindView(R.id.floating_btn) FloatingActionButton floatingBtn;
-    @BindView(R.id.tab_layout) TabLayout tabLayout;
-    @BindView(R.id.account_locked) ImageView accountLockedView;
-    @BindView(R.id.activity_account) View container;
+    private CircularImageView avatar;
+    private ImageView header;
+    private FloatingActionButton floatingBtn;
+    private Button followBtn;
+    private TextView followsYouView;
+    private TabLayout tabLayout;
+    private ImageView accountLockedView;
+    private View container;
+    private TextView followersTextView;
+    private TextView followingTextView;
+    private TextView statusesTextView;
+
+    private boolean hideFab;
+    private int oldOffset;
 
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_account);
-        ButterKnife.bind(this);
+
+        avatar = findViewById(R.id.account_avatar);
+        header = findViewById(R.id.account_header);
+        floatingBtn = findViewById(R.id.floating_btn);
+        followBtn = findViewById(R.id.follow_btn);
+        followsYouView = findViewById(R.id.account_follows_you);
+        tabLayout = findViewById(R.id.tab_layout);
+        accountLockedView = findViewById(R.id.account_locked);
+        container = findViewById(R.id.activity_account);
+        followersTextView = findViewById(R.id.followers_tv);
+        followingTextView = findViewById(R.id.following_tv);
+        statusesTextView = findViewById(R.id.statuses_btn);
 
         if (savedInstanceState != null) {
             accountId = savedInstanceState.getString("accountId");
@@ -108,11 +128,8 @@ public class AccountActivity extends BaseActivity {
         }
         loadedAccount = null;
 
-        SharedPreferences preferences = getPrivatePreferences();
-        String loggedInAccountId = preferences.getString("loggedInAccountId", null);
-
         // Setup the toolbar.
-        final Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
+        final Toolbar toolbar = findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
         ActionBar actionBar = getSupportActionBar();
         if (actionBar != null) {
@@ -121,12 +138,14 @@ public class AccountActivity extends BaseActivity {
             actionBar.setDisplayShowHomeEnabled(true);
         }
 
+        hideFab = PreferenceManager.getDefaultSharedPreferences(this).getBoolean("fabHide", false);
+
         // Add a listener to change the toolbar icon color when it enters/exits its collapsed state.
-        AppBarLayout appBarLayout = (AppBarLayout) findViewById(R.id.account_app_bar_layout);
-        final CollapsingToolbarLayout collapsingToolbar =
-                (CollapsingToolbarLayout) findViewById(R.id.collapsing_toolbar);
+        AppBarLayout appBarLayout = findViewById(R.id.account_app_bar_layout);
+        final CollapsingToolbarLayout collapsingToolbar = findViewById(R.id.collapsing_toolbar);
         appBarLayout.addOnOffsetChangedListener(new AppBarLayout.OnOffsetChangedListener() {
-            @AttrRes int priorAttribute = R.attr.account_toolbar_icon_tint_uncollapsed;
+            @AttrRes
+            int priorAttribute = R.attr.account_toolbar_icon_tint_uncollapsed;
 
             @Override
             public void onOffsetChanged(AppBarLayout appBarLayout, int verticalOffset) {
@@ -134,10 +153,10 @@ public class AccountActivity extends BaseActivity {
                 if (collapsingToolbar.getHeight() + verticalOffset
                         < 2 * ViewCompat.getMinimumHeight(collapsingToolbar)) {
 
-                        toolbar.setTitleTextColor(ThemeUtils.getColor(AccountActivity.this,
-                                android.R.attr.textColorPrimary));
-                        toolbar.setSubtitleTextColor(ThemeUtils.getColor(AccountActivity.this,
-                                android.R.attr.textColorSecondary));
+                    toolbar.setTitleTextColor(ThemeUtils.getColor(AccountActivity.this,
+                            android.R.attr.textColorPrimary));
+                    toolbar.setSubtitleTextColor(ThemeUtils.getColor(AccountActivity.this,
+                            android.R.attr.textColorSecondary));
 
                     attribute = R.attr.account_toolbar_icon_tint_collapsed;
                 } else {
@@ -152,47 +171,82 @@ public class AccountActivity extends BaseActivity {
                     ThemeUtils.setDrawableTint(context, toolbar.getNavigationIcon(), attribute);
                     ThemeUtils.setDrawableTint(context, toolbar.getOverflowIcon(), attribute);
                 }
+
+                if (floatingBtn != null && hideFab && !isSelf && !blocking) {
+                    if (verticalOffset > oldOffset) {
+                        floatingBtn.show();
+                    }
+                    if (verticalOffset < oldOffset) {
+                        floatingBtn.hide();
+                    }
+                }
+                oldOffset = verticalOffset;
             }
         });
 
         // Initialise the default UI states.
         floatingBtn.hide();
+        followBtn.setVisibility(View.GONE);
+        followsYouView.setVisibility(View.GONE);
 
         // Obtain information to fill out the profile.
         obtainAccount();
-        if (!accountId.equals(loggedInAccountId)) {
+
+        AccountEntity activeAccount = TuskyApplication.getAccountManager().getActiveAccount();
+
+        if (accountId.equals(activeAccount.getAccountId())) {
+            isSelf = true;
+        } else {
             isSelf = false;
             obtainRelationships();
-        } else {
-            /* Cause the options menu to update and instead show an options menu for when the
-             * account being shown is their own account. */
-            isSelf = true;
-            invalidateOptionsMenu();
         }
 
         // Setup the tabs and timeline pager.
-        AccountPagerAdapter adapter = new AccountPagerAdapter(getSupportFragmentManager(), this,
+        AccountPagerAdapter adapter = new AccountPagerAdapter(getSupportFragmentManager(),
                 accountId);
         String[] pageTitles = {
-            getString(R.string.title_statuses),
-            getString(R.string.title_follows),
-            getString(R.string.title_followers)
+                getString(R.string.title_statuses),
+                getString(R.string.title_media)
         };
         adapter.setPageTitles(pageTitles);
-        ViewPager viewPager = (ViewPager) findViewById(R.id.pager);
+        final ViewPager viewPager = findViewById(R.id.pager);
         int pageMargin = getResources().getDimensionPixelSize(R.dimen.tab_page_margin);
         viewPager.setPageMargin(pageMargin);
         Drawable pageMarginDrawable = ThemeUtils.getDrawable(this, R.attr.tab_page_margin_drawable,
                 R.drawable.tab_page_margin_dark);
         viewPager.setPageMarginDrawable(pageMarginDrawable);
         viewPager.setAdapter(adapter);
+        viewPager.setOffscreenPageLimit(0);
         tabLayout.setupWithViewPager(viewPager);
-        for (int i = 0; i < tabLayout.getTabCount(); i++) {
-            TabLayout.Tab tab = tabLayout.getTabAt(i);
-            if (tab != null) {
-                tab.setCustomView(adapter.getTabView(i, tabLayout));
+
+        View.OnClickListener accountListClickListener = v -> {
+            AccountListActivity.Type type;
+            switch (v.getId()) {
+                case R.id.followers_tv:
+                    type = AccountListActivity.Type.FOLLOWERS;
+                    break;
+                case R.id.following_tv:
+                    type = AccountListActivity.Type.FOLLOWING;
+                    break;
+                default:
+                    throw new AssertionError();
             }
-        }
+            Intent intent = AccountListActivity.newIntent(AccountActivity.this, type,
+                    accountId);
+            startActivity(intent);
+        };
+        followersTextView.setOnClickListener(accountListClickListener);
+        followingTextView.setOnClickListener(accountListClickListener);
+
+        statusesTextView.setOnClickListener(v -> {
+            // Make nice ripple effect on tab
+
+            //noinspection ConstantConditions
+            tabLayout.getTabAt(0).select();
+            final View poorTabView = ((ViewGroup) tabLayout.getChildAt(0)).getChildAt(0);
+            poorTabView.setPressed(true);
+            tabLayout.postDelayed(() -> poorTabView.setPressed(false), 300);
+        });
     }
 
     @Override
@@ -205,9 +259,10 @@ public class AccountActivity extends BaseActivity {
     }
 
     private void obtainAccount() {
-        mastodonAPI.account(accountId).enqueue(new Callback<Account>() {
+        mastodonApi.account(accountId).enqueue(new Callback<Account>() {
             @Override
-            public void onResponse(Call<Account> call, Response<Account> response) {
+            public void onResponse(@NonNull Call<Account> call,
+                                   @NonNull Response<Account> response) {
                 if (response.isSuccessful()) {
                     onObtainAccountSuccess(response.body());
                 } else {
@@ -216,7 +271,7 @@ public class AccountActivity extends BaseActivity {
             }
 
             @Override
-            public void onFailure(Call<Account> call, Throwable t) {
+            public void onFailure(@NonNull Call<Account> call, @NonNull Throwable t) {
                 onObtainAccountFailure();
             }
         });
@@ -225,9 +280,9 @@ public class AccountActivity extends BaseActivity {
     private void onObtainAccountSuccess(Account account) {
         loadedAccount = account;
 
-        TextView username = (TextView) findViewById(R.id.account_username);
-        TextView displayName = (TextView) findViewById(R.id.account_display_name);
-        TextView note = (TextView) findViewById(R.id.account_note);
+        TextView username = findViewById(R.id.account_username);
+        TextView displayName = findViewById(R.id.account_display_name);
+        TextView note = findViewById(R.id.account_note);
 
         String usernameFormatted = String.format(
                 getString(R.string.status_username_format), account.username);
@@ -241,12 +296,9 @@ public class AccountActivity extends BaseActivity {
             String subtitle = String.format(getString(R.string.status_username_format),
                     account.username);
             getSupportActionBar().setSubtitle(subtitle);
-
         }
 
-        boolean useCustomTabs = PreferenceManager.getDefaultSharedPreferences(this)
-                .getBoolean("customTabs", true);
-        LinkHelper.setClickableText(note, account.note, null, useCustomTabs, new LinkListener() {
+        LinkHelper.setClickableText(note, account.note, null, new LinkListener() {
             @Override
             public void onViewTag(String tag) {
                 Intent intent = new Intent(AccountActivity.this, ViewTagActivity.class);
@@ -271,102 +323,82 @@ public class AccountActivity extends BaseActivity {
         Picasso.with(this)
                 .load(account.avatar)
                 .placeholder(R.drawable.avatar_default)
-                .error(R.drawable.avatar_error)
                 .into(avatar);
         Picasso.with(this)
                 .load(account.header)
                 .placeholder(R.drawable.account_header_default)
                 .into(header);
 
-        NumberFormat nf = NumberFormat.getInstance();
+        long followersCount = Long.parseLong(account.followersCount);
+        long followingCount = Long.parseLong(account.followingCount);
+        long statusesCount = Long.parseLong(account.statusesCount);
+        followersTextView.setText(getString(R.string.title_x_followers, followersCount));
+        followingTextView.setText(getString(R.string.title_x_following, followingCount));
+        statusesTextView.setText(getString(R.string.title_x_statuses, statusesCount));
 
-        // Add counts to the tabs in the TabLayout.
-        String[] counts = {
-            nf.format(Integer.parseInt(account.statusesCount)),
-            nf.format(Integer.parseInt(account.followingCount)),
-            nf.format(Integer.parseInt(account.followersCount)),
-        };
-
-        for (int i = 0; i < tabLayout.getTabCount(); i++) {
-            TabLayout.Tab tab = tabLayout.getTabAt(i);
-            if (tab != null) {
-                View view = tab.getCustomView();
-                if (view != null) {
-                    TextView total = (TextView) view.findViewById(R.id.total);
-                    total.setText(counts[i]);
-                }
-            }
-        }
     }
 
     private void onObtainAccountFailure() {
         Snackbar.make(tabLayout, R.string.error_generic, Snackbar.LENGTH_LONG)
-                .setAction(R.string.action_retry, new View.OnClickListener() {
-                    @Override
-                    public void onClick(View v) {
-                        obtainAccount();
-                    }
-                })
+                .setAction(R.string.action_retry, v -> obtainAccount())
                 .show();
     }
 
     private void obtainRelationships() {
         List<String> ids = new ArrayList<>(1);
         ids.add(accountId);
-        mastodonAPI.relationships(ids).enqueue(new Callback<List<Relationship>>() {
+        mastodonApi.relationships(ids).enqueue(new Callback<List<Relationship>>() {
             @Override
-            public void onResponse(Call<List<Relationship>> call,
-                                   Response<List<Relationship>> response) {
-                if (response.isSuccessful()) {
-                    Relationship relationship = response.body().get(0);
-                    onObtainRelationshipsSuccess(relationship.requested, relationship.following,
-                            relationship.blocking, relationship.muting);
+            public void onResponse(@NonNull Call<List<Relationship>> call,
+                                   @NonNull Response<List<Relationship>> response) {
+                List<Relationship> relationships = response.body();
+                if (response.isSuccessful() && relationships != null) {
+                    Relationship relationship = relationships.get(0);
+                    onObtainRelationshipsSuccess(relationship);
                 } else {
                     onObtainRelationshipsFailure(new Exception(response.message()));
                 }
             }
 
             @Override
-            public void onFailure(Call<List<Relationship>> call, Throwable t) {
+            public void onFailure(@NonNull Call<List<Relationship>> call, @NonNull Throwable t) {
                 onObtainRelationshipsFailure((Exception) t);
             }
         });
     }
 
-    private void onObtainRelationshipsSuccess(boolean followRequested, boolean following,
-                                              boolean blocking, boolean muting) {
-        if (following) {
+    private void onObtainRelationshipsSuccess(Relationship relation) {
+        if (relation.following) {
             followState = FollowState.FOLLOWING;
-        } else if (followRequested) {
+        } else if (relation.requested) {
             followState = FollowState.REQUESTED;
         } else {
             followState = FollowState.NOT_FOLLOWING;
         }
-        this.blocking = blocking;
-        this.muting = muting;
+        this.blocking = relation.blocking;
+        this.muting = relation.muting;
 
-        if (followState != FollowState.NOT_FOLLOWING || !blocking || !muting) {
-            invalidateOptionsMenu();
+        if (relation.followedBy) {
+            followsYouView.setVisibility(View.VISIBLE);
+        } else {
+            followsYouView.setVisibility(View.GONE);
         }
 
         updateButtons();
     }
 
-    private void updateFollowButton(FloatingActionButton button) {
+    private void updateFollowButton(Button button) {
         switch (followState) {
             case NOT_FOLLOWING: {
-                button.setImageResource(R.drawable.ic_person_add_24dp);
-                button.setContentDescription(getString(R.string.action_follow));
+                button.setText(R.string.action_follow);
                 break;
             }
             case REQUESTED: {
-                button.setImageResource(R.drawable.ic_hourglass_24dp);
-                button.setContentDescription(getString(R.string.state_follow_requested));
+                button.setText(R.string.state_follow_requested);
                 break;
             }
             case FOLLOWING: {
-                button.setImageResource(R.drawable.ic_person_minus_24px);
-                button.setContentDescription(getString(R.string.action_unfollow));
+                button.setText(R.string.action_unfollow);
                 break;
             }
         }
@@ -375,22 +407,34 @@ public class AccountActivity extends BaseActivity {
     private void updateButtons() {
         invalidateOptionsMenu();
 
-        if(!isSelf && !blocking) {
+        if (!isSelf && !blocking) {
             floatingBtn.show();
+            followBtn.setVisibility(View.VISIBLE);
 
-            updateFollowButton(floatingBtn);
+            updateFollowButton(followBtn);
 
-            floatingBtn.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-                    if (followState != FollowState.REQUESTED) {
+            floatingBtn.setOnClickListener(v -> mention());
+
+            followBtn.setOnClickListener(v -> {
+                switch (followState) {
+                    case NOT_FOLLOWING: {
                         follow(accountId);
-                    } else {
-                        showFollowRequestPendingDialog(accountId);
+                        break;
                     }
-                    updateFollowButton(floatingBtn);
+                    case REQUESTED: {
+                        showFollowRequestPendingDialog();
+                        break;
+                    }
+                    case FOLLOWING: {
+                        showUnfollowWarningDialog();
+                        break;
+                    }
                 }
+                updateFollowButton(followBtn);
             });
+        } else {
+            floatingBtn.hide();
+            followBtn.setVisibility(View.GONE);
         }
     }
 
@@ -407,9 +451,11 @@ public class AccountActivity extends BaseActivity {
     private String getFollowAction() {
         switch (followState) {
             default:
-            case NOT_FOLLOWING: return getString(R.string.action_follow);
+            case NOT_FOLLOWING:
+                return getString(R.string.action_follow);
             case REQUESTED:
-            case FOLLOWING: return getString(R.string.action_unfollow);
+            case FOLLOWING:
+                return getString(R.string.action_unfollow);
         }
     }
 
@@ -447,9 +493,10 @@ public class AccountActivity extends BaseActivity {
     private void follow(final String id) {
         Callback<Relationship> cb = new Callback<Relationship>() {
             @Override
-            public void onResponse(Call<Relationship> call, Response<Relationship> response) {
-                if (response.isSuccessful()) {
-                    Relationship relationship = response.body();
+            public void onResponse(@NonNull Call<Relationship> call,
+                                   @NonNull Response<Relationship> response) {
+                Relationship relationship = response.body();
+                if (response.isSuccessful() && relationship != null) {
                     if (relationship.following) {
                         followState = FollowState.FOLLOWING;
                     } else if (relationship.requested) {
@@ -467,50 +514,56 @@ public class AccountActivity extends BaseActivity {
             }
 
             @Override
-            public void onFailure(Call<Relationship> call, Throwable t) {
+            public void onFailure(@NonNull Call<Relationship> call, @NonNull Throwable t) {
                 onFollowFailure(id);
             }
         };
 
         Assert.expect(followState != FollowState.REQUESTED);
         switch (followState) {
-            case NOT_FOLLOWING: { mastodonAPI.followAccount(id).enqueue(cb);   break; }
-            case FOLLOWING:     { mastodonAPI.unfollowAccount(id).enqueue(cb); break; }
+            case NOT_FOLLOWING: {
+                mastodonApi.followAccount(id).enqueue(cb);
+                break;
+            }
+            case FOLLOWING: {
+                mastodonApi.unfollowAccount(id).enqueue(cb);
+                break;
+            }
         }
     }
 
     private void onFollowFailure(final String id) {
-        View.OnClickListener listener = new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                follow(id);
-            }
-        };
+        View.OnClickListener listener = v -> follow(id);
         Snackbar.make(container, R.string.error_generic, Snackbar.LENGTH_LONG)
                 .setAction(R.string.action_retry, listener)
                 .show();
     }
 
-    private void showFollowRequestPendingDialog(final String id) {
-        DialogInterface.OnClickListener waitListener = new DialogInterface.OnClickListener() {
-            @Override
-            public void onClick(DialogInterface dialog, int which) {
-                dialog.dismiss();
-            }
-        };
-        new AlertDialog.Builder(AccountActivity.this)
+    private void showFollowRequestPendingDialog() {
+        new AlertDialog.Builder(this)
                 .setMessage(R.string.dialog_message_follow_request)
-                .setPositiveButton(R.string.action_ok, waitListener)
+                .setPositiveButton(android.R.string.ok, null)
+                .show();
+    }
+
+    private void showUnfollowWarningDialog() {
+        DialogInterface.OnClickListener unfollowListener = (dialogInterface, i) -> follow(accountId);
+        new AlertDialog.Builder(this)
+                .setMessage(R.string.dialog_unfollow_warning)
+                .setPositiveButton(android.R.string.ok, unfollowListener)
+                .setNegativeButton(android.R.string.cancel, null)
                 .show();
     }
 
     private void block(final String id) {
         Callback<Relationship> cb = new Callback<Relationship>() {
             @Override
-            public void onResponse(Call<Relationship> call, Response<Relationship> response) {
-                if (response.isSuccessful()) {
+            public void onResponse(@NonNull Call<Relationship> call,
+                                   @NonNull Response<Relationship> response) {
+                Relationship relationship = response.body();
+                if (response.isSuccessful() && relationship != null) {
                     broadcast(TimelineReceiver.Types.BLOCK_ACCOUNT, id);
-                    blocking = response.body().blocking;
+                    blocking = relationship.blocking;
                     updateButtons();
                 } else {
                     onBlockFailure(id);
@@ -518,37 +571,33 @@ public class AccountActivity extends BaseActivity {
             }
 
             @Override
-            public void onFailure(Call<Relationship> call, Throwable t) {
+            public void onFailure(@NonNull Call<Relationship> call, @NonNull Throwable t) {
                 onBlockFailure(id);
             }
         };
         if (blocking) {
-            mastodonAPI.unblockAccount(id).enqueue(cb);
+            mastodonApi.unblockAccount(id).enqueue(cb);
         } else {
-            mastodonAPI.blockAccount(id).enqueue(cb);
+            mastodonApi.blockAccount(id).enqueue(cb);
         }
     }
 
     private void onBlockFailure(final String id) {
-        View.OnClickListener listener = new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                block(id);
-            }
-        };
+        View.OnClickListener listener = v -> block(id);
         Snackbar.make(container, R.string.error_generic, Snackbar.LENGTH_LONG)
                 .setAction(R.string.action_retry, listener)
                 .show();
     }
 
-
     private void mute(final String id) {
         Callback<Relationship> cb = new Callback<Relationship>() {
             @Override
-            public void onResponse(Call<Relationship> call, Response<Relationship> response) {
-                if (response.isSuccessful()) {
+            public void onResponse(@NonNull Call<Relationship> call,
+                                   @NonNull Response<Relationship> response) {
+                Relationship relationship = response.body();
+                if (response.isSuccessful() && relationship != null) {
                     broadcast(TimelineReceiver.Types.MUTE_ACCOUNT, id);
-                    muting = response.body().muting;
+                    muting = relationship.muting;
                     updateButtons();
                 } else {
                     onMuteFailure(id);
@@ -556,28 +605,35 @@ public class AccountActivity extends BaseActivity {
             }
 
             @Override
-            public void onFailure(Call<Relationship> call, Throwable t) {
+            public void onFailure(@NonNull Call<Relationship> call, @NonNull Throwable t) {
                 onMuteFailure(id);
             }
         };
 
         if (muting) {
-            mastodonAPI.unmuteAccount(id).enqueue(cb);
+            mastodonApi.unmuteAccount(id).enqueue(cb);
         } else {
-            mastodonAPI.muteAccount(id).enqueue(cb);
+            mastodonApi.muteAccount(id).enqueue(cb);
         }
     }
 
     private void onMuteFailure(final String id) {
-        View.OnClickListener listener = new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                mute(id);
-            }
-        };
+        View.OnClickListener listener = v -> mute(id);
         Snackbar.make(container, R.string.error_generic, Snackbar.LENGTH_LONG)
                 .setAction(R.string.action_retry, listener)
                 .show();
+    }
+
+    private boolean mention() {
+        if (loadedAccount == null) {
+            // If the account isn't loaded yet, eat the input.
+            return false;
+        }
+        Intent intent = new ComposeActivity.IntentBuilder()
+                .mentionedUsernames(Collections.singleton(loadedAccount.username))
+                .build(this);
+        startActivity(intent);
+        return true;
     }
 
     private void broadcast(String action, String id) {
@@ -594,23 +650,14 @@ public class AccountActivity extends BaseActivity {
                 return true;
             }
             case R.id.action_mention: {
-                if (loadedAccount == null) {
-                    // If the account isn't loaded yet, eat the input.
-                    return false;
-                }
-                Intent intent = new Intent(this, ComposeActivity.class);
-                intent.putExtra("mentioned_usernames", new String[] { loadedAccount.username });
-                startActivity(intent);
-                return true;
+                return mention();
             }
             case R.id.action_open_in_web: {
                 if (loadedAccount == null) {
                     // If the account isn't loaded yet, eat the input.
                     return false;
                 }
-                Uri uri = Uri.parse(loadedAccount.url);
-                Intent intent = new Intent(Intent.ACTION_VIEW, uri);
-                startActivity(intent);
+                LinkHelper.openLink(loadedAccount.url, this);
                 return true;
             }
             case R.id.action_follow: {
@@ -628,4 +675,14 @@ public class AccountActivity extends BaseActivity {
         }
         return super.onOptionsItemSelected(item);
     }
+
+    @Nullable
+    @Override
+    public FloatingActionButton getActionButton() {
+        if (!isSelf && !blocking) {
+            return floatingBtn;
+        }
+        return null;
+    }
+
 }
