@@ -15,15 +15,15 @@
 
 package com.keylesspalace.tusky;
 
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.content.res.Configuration;
-import android.graphics.Typeface;
+import android.graphics.Color;
 import android.graphics.drawable.Drawable;
 import android.net.Uri;
 import android.os.Bundle;
-import android.os.PersistableBundle;
+import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
 import android.support.design.widget.FloatingActionButton;
 import android.support.design.widget.TabLayout;
 import android.support.graphics.drawable.VectorDrawableCompat;
@@ -31,21 +31,18 @@ import android.support.v4.content.ContextCompat;
 import android.support.v4.content.LocalBroadcastManager;
 import android.support.v4.view.ViewPager;
 import android.support.v7.app.AlertDialog;
-import android.text.SpannableStringBuilder;
-import android.text.Spanned;
-import android.text.TextUtils;
-import android.text.style.StyleSpan;
 import android.util.Log;
-import android.view.View;
+import android.view.KeyEvent;
+import android.widget.ImageButton;
 import android.widget.ImageView;
-import android.widget.TextView;
 
-import com.arlib.floatingsearchview.FloatingSearchView;
-import com.arlib.floatingsearchview.suggestions.SearchSuggestionsAdapter;
-import com.arlib.floatingsearchview.suggestions.model.SearchSuggestion;
+import com.keylesspalace.tusky.db.AccountEntity;
+import com.keylesspalace.tusky.db.AccountManager;
 import com.keylesspalace.tusky.entity.Account;
+import com.keylesspalace.tusky.interfaces.ActionButtonActivity;
 import com.keylesspalace.tusky.pager.TimelinePagerAdapter;
 import com.keylesspalace.tusky.receiver.TimelineReceiver;
+import com.keylesspalace.tusky.util.NotificationHelper;
 import com.keylesspalace.tusky.util.ThemeUtils;
 import com.mikepenz.google_material_typeface_library.GoogleMaterial;
 import com.mikepenz.materialdrawer.AccountHeader;
@@ -55,6 +52,7 @@ import com.mikepenz.materialdrawer.DrawerBuilder;
 import com.mikepenz.materialdrawer.model.DividerDrawerItem;
 import com.mikepenz.materialdrawer.model.PrimaryDrawerItem;
 import com.mikepenz.materialdrawer.model.ProfileDrawerItem;
+import com.mikepenz.materialdrawer.model.ProfileSettingDrawerItem;
 import com.mikepenz.materialdrawer.model.SecondaryDrawerItem;
 import com.mikepenz.materialdrawer.model.interfaces.IDrawerItem;
 import com.mikepenz.materialdrawer.model.interfaces.IProfile;
@@ -64,58 +62,75 @@ import com.squareup.picasso.Picasso;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Stack;
 
-import butterknife.BindView;
-import butterknife.ButterKnife;
 import retrofit2.Call;
 import retrofit2.Callback;
 import retrofit2.Response;
 
-public class MainActivity extends BaseActivity {
+public class MainActivity extends BaseActivity implements ActionButtonActivity {
     private static final String TAG = "MainActivity"; // logging tag
-    protected static int COMPOSE_RESULT = 1;
+    private static final long DRAWER_ITEM_ADD_ACCOUNT = -13;
+    private static final long DRAWER_ITEM_EDIT_PROFILE = 0;
+    private static final long DRAWER_ITEM_FAVOURITES = 1;
+    private static final long DRAWER_ITEM_MUTED_USERS = 2;
+    private static final long DRAWER_ITEM_BLOCKED_USERS = 3;
+    private static final long DRAWER_ITEM_SEARCH = 4;
+    private static final long DRAWER_ITEM_PREFERENCES = 5;
+    private static final long DRAWER_ITEM_ABOUT = 6;
+    private static final long DRAWER_ITEM_LOG_OUT = 7;
+    private static final long DRAWER_ITEM_FOLLOW_REQUESTS = 8;
+    private static final long DRAWER_ITEM_SAVED_TOOT = 9;
+    private static final long DRAWER_ITEM_LISTS = 10;
 
-    private String loggedInAccountId;
-    private String loggedInAccountUsername;
-    private Stack<Integer> pageHistory;
+    private static int COMPOSE_RESULT = 1;
+
+    private FloatingActionButton composeButton;
     private AccountHeader headerResult;
     private Drawer drawer;
-
-    @BindView(R.id.floating_search_view) FloatingSearchView searchView;
-    @BindView(R.id.floating_btn) FloatingActionButton floatingBtn;
-    @BindView(R.id.tab_layout) TabLayout tabLayout;
-    @BindView(R.id.pager) ViewPager viewPager;
-
-    public FloatingActionButton composeButton;
+    private ViewPager viewPager;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
 
-        pageHistory = new Stack<>();
-        if (savedInstanceState != null) {
-            List<Integer> restoredHistory = savedInstanceState.getIntegerArrayList("pageHistory");
-            if (restoredHistory != null) {
-                pageHistory.addAll(restoredHistory);
+        // account switching has to be done before MastodonApi is created in super.onCreate
+        Intent intent = getIntent();
+
+        int tabPosition = 0;
+
+        if (intent != null) {
+            long accountId = intent.getLongExtra(NotificationHelper.ACCOUNT_ID, -1);
+
+            if(accountId != -1) {
+                // user clicked a notification, show notification tab and switch user if necessary
+                tabPosition = 1;
+                AccountEntity account = TuskyApplication.getAccountManager().getActiveAccount();
+
+                if (account == null || accountId != account.getId()) {
+                    TuskyApplication.getAccountManager().setActiveAccount(accountId);
+                }
             }
         }
 
-        ButterKnife.bind(this);
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
 
-        floatingBtn.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(getApplicationContext(), ComposeActivity.class);
-                startActivityForResult(intent, COMPOSE_RESULT);
-            }
+        FloatingActionButton floatingBtn = findViewById(R.id.floating_btn);
+        ImageButton drawerToggle = findViewById(R.id.drawer_toggle);
+        TabLayout tabLayout = findViewById(R.id.tab_layout);
+        viewPager = findViewById(R.id.pager);
+
+        floatingBtn.setOnClickListener(v -> {
+            Intent composeIntent = new Intent(getApplicationContext(), ComposeActivity.class);
+            startActivityForResult(composeIntent, COMPOSE_RESULT);
         });
 
         setupDrawer();
-        setupSearchView();
 
-        /* Fetch user info while we're doing other things. This has to be after setting up the
+        // Setup the navigation drawer toggle button.
+        ThemeUtils.setDrawableTint(this, drawerToggle.getDrawable(), R.attr.toolbar_icon_tint);
+        drawerToggle.setOnClickListener(v -> drawer.openDrawer());
+
+        /* Fetch user info while we're doing other things. This has to be done after setting up the
          * drawer, though, because its callback touches the header in the drawer. */
         fetchUserInfo();
 
@@ -149,21 +164,25 @@ public class MainActivity extends BaseActivity {
             tab.setContentDescription(pageTitles[i]);
         }
 
+        if (tabPosition != 0) {
+            TabLayout.Tab tab = tabLayout.getTabAt(tabPosition);
+            if (tab != null) {
+                tab.select();
+            } else {
+                tabPosition = 0;
+            }
+        }
+
         tabLayout.addOnTabSelectedListener(new TabLayout.OnTabSelectedListener() {
             @Override
             public void onTabSelected(TabLayout.Tab tab) {
                 viewPager.setCurrentItem(tab.getPosition());
 
-                if (pageHistory.isEmpty()) {
-                    pageHistory.push(0);
-                }
-
-                if (pageHistory.contains(tab.getPosition())) {
-                    pageHistory.remove(pageHistory.indexOf(tab.getPosition()));
-                }
-
-                pageHistory.push(tab.getPosition());
                 tintTab(tab, true);
+
+                if(tab.getPosition() == 1) {
+                    NotificationHelper.clearNotificationsForActiveAccount(MainActivity.this);
+                }
             }
 
             @Override
@@ -172,28 +191,15 @@ public class MainActivity extends BaseActivity {
             }
 
             @Override
-            public void onTabReselected(TabLayout.Tab tab) {}
+            public void onTabReselected(TabLayout.Tab tab) { }
         });
 
-        Intent intent = getIntent();
-
-        int tabSelected = 0;
-        if (intent != null) {
-            int tabPosition = intent.getIntExtra("tab_position", 0);
-            if (tabPosition != 0) {
-                TabLayout.Tab tab = tabLayout.getTabAt(tabPosition);
-                if (tab != null) {
-                    tab.select();
-                    tabSelected = tabPosition;
-                }
-            }
-        }
         for (int i = 0; i < 4; i++) {
-            tintTab(tabLayout.getTabAt(i), i == tabSelected);
+            tintTab(tabLayout.getTabAt(i), i == tabPosition);
         }
 
         // Setup push notifications
-        if (arePushNotificationsEnabled()) {
+        if (NotificationHelper.areNotificationsEnabled(this)) {
             enablePushNotifications();
         } else {
             disablePushNotifications();
@@ -206,13 +212,7 @@ public class MainActivity extends BaseActivity {
     protected void onResume() {
         super.onResume();
 
-        SharedPreferences notificationPreferences = getApplicationContext()
-                .getSharedPreferences("Notifications", MODE_PRIVATE);
-        notificationPreferences.edit()
-                .putString("current", "[]")
-                .apply();
-
-        pushNotificationClient.clearNotifications(this);
+        NotificationHelper.clearNotificationsForActiveAccount(this);
 
         /* After editing a profile, the profile header in the navigation drawer needs to be
          * refreshed */
@@ -223,314 +223,7 @@ public class MainActivity extends BaseActivity {
                     .putBoolean("refreshProfileHeader", false)
                     .apply();
         }
-    }
 
-    @Override
-    public void onSaveInstanceState(Bundle outState, PersistableBundle outPersistentState) {
-        ArrayList<Integer> pageHistoryList = new ArrayList<>();
-        pageHistoryList.addAll(pageHistory);
-        outState.putIntegerArrayList("pageHistory", pageHistoryList);
-        super.onSaveInstanceState(outState, outPersistentState);
-    }
-
-    private void tintTab(TabLayout.Tab tab, boolean tinted) {
-        int color = (tinted) ? R.attr.tab_icon_selected_tint : R.attr.toolbar_icon_tint;
-        ThemeUtils.setDrawableTint(this, tab.getIcon(), color);
-    }
-
-    private void setupDrawer() {
-        headerResult = new AccountHeaderBuilder()
-                .withActivity(this)
-                .withSelectionListEnabledForSingleProfile(false)
-                .withDividerBelowHeader(false)
-                .withOnAccountHeaderProfileImageListener(new AccountHeader.OnAccountHeaderProfileImageListener() {
-                    @Override
-                    public boolean onProfileImageClick(View view, IProfile profile, boolean current) {
-                        if (current && loggedInAccountId != null) {
-                            Intent intent = new Intent(MainActivity.this, AccountActivity.class);
-                            intent.putExtra("id", loggedInAccountId);
-                            startActivity(intent);
-                            return true;
-                        }
-                        return false;
-                    }
-
-                    @Override
-                    public boolean onProfileImageLongClick(View view, IProfile profile, boolean current) {
-                        return false;
-                    }
-                })
-                .withCompactStyle(true)
-                .build();
-        headerResult.getView()
-                .findViewById(R.id.material_drawer_account_header_current)
-                .setContentDescription(getString(R.string.action_view_profile));
-
-        DrawerImageLoader.init(new AbstractDrawerImageLoader() {
-            @Override
-            public void set(ImageView imageView, Uri uri, Drawable placeholder) {
-                Picasso.with(imageView.getContext()).load(uri).placeholder(placeholder).into(imageView);
-            }
-
-            @Override
-            public void cancel(ImageView imageView) {
-                Picasso.with(imageView.getContext()).cancelRequest(imageView);
-            }
-        });
-
-        VectorDrawableCompat muteDrawable = VectorDrawableCompat.create(getResources(),
-                R.drawable.ic_mute_24dp, getTheme());
-        ThemeUtils.setDrawableTint(this, muteDrawable, R.attr.toolbar_icon_tint);
-
-        drawer = new DrawerBuilder()
-                .withActivity(this)
-                //.withToolbar(toolbar)
-                .withAccountHeader(headerResult)
-                .withHasStableIds(true)
-                .withSelectedItem(-1)
-                .addDrawerItems(
-                        new PrimaryDrawerItem().withIdentifier(0).withName(getString(R.string.action_edit_profile)).withSelectable(false).withIcon(GoogleMaterial.Icon.gmd_person),
-                        new PrimaryDrawerItem().withIdentifier(1).withName(getString(R.string.action_view_favourites)).withSelectable(false).withIcon(GoogleMaterial.Icon.gmd_star),
-                        new PrimaryDrawerItem().withIdentifier(2).withName(getString(R.string.action_view_mutes)).withSelectable(false).withIcon(muteDrawable),
-                        new PrimaryDrawerItem().withIdentifier(3).withName(getString(R.string.action_view_blocks)).withSelectable(false).withIcon(GoogleMaterial.Icon.gmd_block),
-                        new DividerDrawerItem(),
-                        new SecondaryDrawerItem().withIdentifier(4).withName(getString(R.string.action_view_preferences)).withSelectable(false).withIcon(GoogleMaterial.Icon.gmd_settings),
-                        new SecondaryDrawerItem().withIdentifier(5).withName(getString(R.string.about_title_activity)).withSelectable(false).withIcon(GoogleMaterial.Icon.gmd_info),
-                        new SecondaryDrawerItem().withIdentifier(6).withName(getString(R.string.action_logout)).withSelectable(false).withIcon(GoogleMaterial.Icon.gmd_exit_to_app)
-                )
-                .withOnDrawerItemClickListener(new Drawer.OnDrawerItemClickListener() {
-                    @Override
-                    public boolean onItemClick(View view, int position, IDrawerItem drawerItem) {
-                        if (drawerItem != null) {
-                            long drawerItemIdentifier = drawerItem.getIdentifier();
-
-                            if (drawerItemIdentifier == 0) {
-                                Intent intent = new Intent(MainActivity.this, EditProfileActivity.class);
-                                startActivity(intent);
-                            } else if (drawerItemIdentifier == 1) {
-                                Intent intent = new Intent(MainActivity.this, FavouritesActivity.class);
-                                startActivity(intent);
-                            } else if (drawerItemIdentifier == 2) {
-                                Intent intent = new Intent(MainActivity.this, AccountListActivity.class);
-                                intent.putExtra("type", AccountListActivity.Type.MUTES);
-                                startActivity(intent);
-                            } else if (drawerItemIdentifier == 3) {
-                                Intent intent = new Intent(MainActivity.this, AccountListActivity.class);
-                                intent.putExtra("type", AccountListActivity.Type.BLOCKS);
-                                startActivity(intent);
-                            } else if (drawerItemIdentifier == 4) {
-                                Intent intent = new Intent(MainActivity.this, PreferencesActivity.class);
-                                startActivity(intent);
-                            } else if (drawerItemIdentifier == 5) {
-                                Intent intent = new Intent(MainActivity.this, AboutActivity.class);
-                                startActivity(intent);
-                            } else if (drawerItemIdentifier == 6) {
-                                logout();
-                            } else if (drawerItemIdentifier == 7) {
-                                Intent intent = new Intent(MainActivity.this, AccountListActivity.class);
-                                intent.putExtra("type", AccountListActivity.Type.FOLLOW_REQUESTS);
-                                startActivity(intent);
-                            }
-                        }
-
-                        return false;
-                    }
-                })
-                .build();
-    }
-
-    private void logout() {
-        new AlertDialog.Builder(this)
-                .setTitle(R.string.action_logout)
-                .setMessage(R.string.action_logout_confirm)
-                .setPositiveButton(android.R.string.yes, new DialogInterface.OnClickListener() {
-                    @Override
-                    public void onClick(DialogInterface dialog, int which) {
-                        if (arePushNotificationsEnabled()) disablePushNotifications();
-
-                        getPrivatePreferences().edit()
-                                .remove("domain")
-                                .remove("accessToken")
-                                .remove("appAccountId")
-                                .apply();
-
-                        Intent intent = new Intent(MainActivity.this, LoginActivity.class);
-                        startActivity(intent);
-                        finish();
-                    }
-                })
-                .setNegativeButton(android.R.string.no, null)
-                .show();
-    }
-
-    private void setupSearchView() {
-        searchView.attachNavigationDrawerToMenuButton(drawer.getDrawerLayout());
-
-        // Setup content descriptions for the different elements in the search view.
-        final View leftAction = searchView.findViewById(R.id.left_action);
-        leftAction.setContentDescription(getString(R.string.action_open_drawer));
-        searchView.setOnFocusChangeListener(new FloatingSearchView.OnFocusChangeListener() {
-            @Override
-            public void onFocus() {
-                leftAction.setContentDescription(getString(R.string.action_close));
-            }
-
-            @Override
-            public void onFocusCleared() {
-                leftAction.setContentDescription(getString(R.string.action_open_drawer));
-            }
-        });
-        View clearButton = searchView.findViewById(R.id.clear_btn);
-        clearButton.setContentDescription(getString(R.string.action_clear));
-
-        searchView.setOnQueryChangeListener(new FloatingSearchView.OnQueryChangeListener() {
-            @Override
-            public void onSearchTextChanged(String oldQuery, String newQuery) {
-                if (!oldQuery.equals("") && newQuery.equals("")) {
-                    searchView.clearSuggestions();
-                    return;
-                }
-
-                if (newQuery.length() < 3) {
-                    return;
-                }
-
-                searchView.showProgress();
-
-                mastodonAPI.searchAccounts(newQuery, false, 5).enqueue(new Callback<List<Account>>() {
-                    @Override
-                    public void onResponse(Call<List<Account>> call, Response<List<Account>> response) {
-                        if (response.isSuccessful()) {
-                            searchView.swapSuggestions(response.body());
-                            searchView.hideProgress();
-                        } else {
-                            searchView.hideProgress();
-                        }
-                    }
-
-                    @Override
-                    public void onFailure(Call<List<Account>> call, Throwable t) {
-                        searchView.hideProgress();
-                    }
-                });
-            }
-        });
-
-        searchView.setOnSearchListener(new FloatingSearchView.OnSearchListener() {
-            @Override
-            public void onSuggestionClicked(SearchSuggestion searchSuggestion) {
-                Account accountSuggestion = (Account) searchSuggestion;
-                Intent intent = new Intent(MainActivity.this, AccountActivity.class);
-                intent.putExtra("id", accountSuggestion.id);
-                startActivity(intent);
-            }
-
-            @Override
-            public void onSearchAction(String currentQuery) {}
-        });
-
-        searchView.setOnBindSuggestionCallback(new SearchSuggestionsAdapter.OnBindSuggestionCallback() {
-            @Override
-            public void onBindSuggestion(View suggestionView, ImageView leftIcon, TextView textView, SearchSuggestion item, int itemPosition) {
-                Account accountSuggestion = ((Account) item);
-
-                Picasso.with(MainActivity.this)
-                        .load(accountSuggestion.avatar)
-                        .placeholder(R.drawable.avatar_default)
-                        .into(leftIcon);
-
-                String searchStr = accountSuggestion.getDisplayName() + " " + accountSuggestion.username;
-                final SpannableStringBuilder str = new SpannableStringBuilder(searchStr);
-
-                str.setSpan(new StyleSpan(Typeface.BOLD), 0, accountSuggestion.getDisplayName().length(), Spanned.SPAN_EXCLUSIVE_EXCLUSIVE);
-                textView.setText(str);
-                textView.setMaxLines(1);
-                textView.setEllipsize(TextUtils.TruncateAt.END);
-            }
-        });
-    }
-
-    private void fetchUserInfo() {
-        SharedPreferences preferences = getPrivatePreferences();
-        final String domain = preferences.getString("domain", null);
-        String id = preferences.getString("loggedInAccountId", null);
-        String username = preferences.getString("loggedInAccountUsername", null);
-
-        if (id != null && username != null) {
-            loggedInAccountId = id;
-            loggedInAccountUsername = username;
-        }
-
-        mastodonAPI.accountVerifyCredentials().enqueue(new Callback<Account>() {
-            @Override
-            public void onResponse(Call<Account> call, Response<Account> response) {
-                if (!response.isSuccessful()) {
-                    onFetchUserInfoFailure(new Exception(response.message()));
-                    return;
-                }
-                onFetchUserInfoSuccess(response.body(), domain);
-            }
-
-            @Override
-            public void onFailure(Call<Account> call, Throwable t) {
-                onFetchUserInfoFailure((Exception) t);
-            }
-        });
-    }
-
-    private void onFetchUserInfoSuccess(Account me, String domain) {
-        // Add the header image and avatar from the account, into the navigation drawer header.
-        headerResult.clear();
-
-        ImageView background = headerResult.getHeaderBackgroundView();
-        int backgroundWidth = background.getWidth();
-        int backgroundHeight = background.getHeight();
-        if (backgroundWidth == 0 || backgroundHeight == 0) {
-            /* The header ImageView may not be layed out when the verify credentials call returns so
-             * measure the dimensions and use those. */
-            background.measure(View.MeasureSpec.EXACTLY, View.MeasureSpec.EXACTLY);
-            backgroundWidth = background.getMeasuredWidth();
-            backgroundHeight = background.getMeasuredHeight();
-        }
-
-        background.setBackgroundColor(ContextCompat.getColor(this, R.color.window_background_dark));
-
-        Picasso.with(MainActivity.this)
-                .load(me.header)
-                .placeholder(R.drawable.account_header_default)
-                .resize(backgroundWidth, backgroundHeight)
-                .centerCrop()
-                .into(background);
-
-        headerResult.addProfiles(
-                new ProfileDrawerItem()
-                        .withName(me.getDisplayName())
-                        .withEmail(String.format("%s@%s", me.username, domain))
-                        .withIcon(me.avatar)
-        );
-
-        // Show follow requests in the menu, if this is a locked account.
-        if (me.locked) {
-            PrimaryDrawerItem followRequestsItem = new PrimaryDrawerItem()
-                    .withIdentifier(7)
-                    .withName(R.string.action_view_follow_requests)
-                    .withSelectable(false)
-                    .withIcon(GoogleMaterial.Icon.gmd_person_add);
-            drawer.addItemAtPosition(followRequestsItem, 3);
-        }
-
-        // Update the current login information.
-        loggedInAccountId = me.id;
-        loggedInAccountUsername = me.username;
-        getPrivatePreferences().edit()
-                .putString("loggedInAccountId", loggedInAccountId)
-                .putString("loggedInAccountUsername", loggedInAccountUsername)
-                .putBoolean("loggedInAccountLocked", me.locked)
-                .apply();
-    }
-
-    private void onFetchUserInfoFailure(Exception exception) {
-        Log.e(TAG, "Failed to fetch user info. " + exception.getMessage());
     }
 
     @Override
@@ -545,19 +238,312 @@ public class MainActivity extends BaseActivity {
 
     @Override
     public void onBackPressed() {
-        if(drawer != null && drawer.isDrawerOpen()) {
+        if (drawer != null && drawer.isDrawerOpen()) {
             drawer.closeDrawer();
-        } else if(pageHistory.size() < 2) {
-            super.onBackPressed();
+        } else if (viewPager.getCurrentItem() != 0) {
+            viewPager.setCurrentItem(0);
         } else {
-            pageHistory.pop();
-            viewPager.setCurrentItem(pageHistory.peek());
+            super.onBackPressed();
         }
+    }
+
+    @Override
+    public boolean onKeyDown(int keyCode, KeyEvent event) {
+        switch (keyCode) {
+            case KeyEvent.KEYCODE_MENU: {
+                if (drawer.isDrawerOpen()) {
+                    drawer.closeDrawer();
+                } else {
+                    drawer.openDrawer();
+                }
+                return true;
+            }
+            case KeyEvent.KEYCODE_SEARCH: {
+                startActivity(new Intent(this, SearchActivity.class));
+                return true;
+            }
+        }
+        return super.onKeyDown(keyCode, event);
     }
 
     // Fix for GitHub issues #190, #259 (MainActivity won't restart on screen rotation.)
     @Override
     public void onConfigurationChanged(Configuration newConfig) {
         super.onConfigurationChanged(newConfig);
+    }
+
+    private void tintTab(TabLayout.Tab tab, boolean tinted) {
+        int color = (tinted) ? R.attr.tab_icon_selected_tint : R.attr.toolbar_icon_tint;
+        ThemeUtils.setDrawableTint(this, tab.getIcon(), color);
+    }
+
+    private void setupDrawer() {
+        headerResult = new AccountHeaderBuilder()
+                .withActivity(this)
+                .withDividerBelowHeader(false)
+                .withHeaderBackgroundScaleType(ImageView.ScaleType.CENTER_CROP)
+                .withCurrentProfileHiddenInList(true)
+                .withOnAccountHeaderListener((view, profile, current) -> handleProfileClick(profile, current))
+                .addProfiles(
+                        new ProfileSettingDrawerItem()
+                                .withIdentifier(DRAWER_ITEM_ADD_ACCOUNT)
+                                .withName(R.string.add_account_name)
+                                .withDescription(R.string.add_account_description)
+                                .withIcon(GoogleMaterial.Icon.gmd_add))
+                .build();
+
+        headerResult.getView()
+                .findViewById(R.id.material_drawer_account_header_current)
+                .setContentDescription(getString(R.string.action_view_profile));
+
+        ImageView background = headerResult.getHeaderBackgroundView();
+        background.setColorFilter(ContextCompat.getColor(this, R.color.header_background_filter));
+        background.setBackgroundColor(ContextCompat.getColor(this, R.color.window_background_dark));
+
+        DrawerImageLoader.init(new AbstractDrawerImageLoader() {
+            @Override
+            public void set(ImageView imageView, Uri uri, Drawable placeholder, String tag) {
+                Picasso.with(imageView.getContext()).load(uri).placeholder(placeholder).into(imageView);
+            }
+
+            @Override
+            public void cancel(ImageView imageView) {
+                Picasso.with(imageView.getContext()).cancelRequest(imageView);
+            }
+        });
+
+        VectorDrawableCompat muteDrawable = VectorDrawableCompat.create(getResources(),
+                R.drawable.ic_mute_24dp, getTheme());
+        ThemeUtils.setDrawableTint(this, muteDrawable, R.attr.toolbar_icon_tint);
+
+        List<IDrawerItem> listItem = new ArrayList<>();
+        listItem.add(new PrimaryDrawerItem().withIdentifier(DRAWER_ITEM_EDIT_PROFILE).withName(getString(R.string.action_edit_profile)).withSelectable(false).withIcon(GoogleMaterial.Icon.gmd_person));
+        listItem.add(new PrimaryDrawerItem().withIdentifier(DRAWER_ITEM_FAVOURITES).withName(getString(R.string.action_view_favourites)).withSelectable(false).withIcon(GoogleMaterial.Icon.gmd_star));
+        listItem.add(new PrimaryDrawerItem().withIdentifier(DRAWER_ITEM_LISTS).withName(R.string.action_lists).withSelectable(false).withIcon(GoogleMaterial.Icon.gmd_list));
+        listItem.add(new PrimaryDrawerItem().withIdentifier(DRAWER_ITEM_MUTED_USERS).withName(getString(R.string.action_view_mutes)).withSelectable(false).withIcon(muteDrawable));
+        listItem.add(new PrimaryDrawerItem().withIdentifier(DRAWER_ITEM_BLOCKED_USERS).withName(getString(R.string.action_view_blocks)).withSelectable(false).withIcon(GoogleMaterial.Icon.gmd_block));
+        listItem.add(new PrimaryDrawerItem().withIdentifier(DRAWER_ITEM_SEARCH).withName(getString(R.string.action_search)).withSelectable(false).withIcon(GoogleMaterial.Icon.gmd_search));
+        listItem.add(new PrimaryDrawerItem().withIdentifier(DRAWER_ITEM_SAVED_TOOT).withName(getString(R.string.action_access_saved_toot)).withSelectable(false).withIcon(GoogleMaterial.Icon.gmd_save));
+        listItem.add(new DividerDrawerItem());
+        listItem.add(new SecondaryDrawerItem().withIdentifier(DRAWER_ITEM_PREFERENCES).withName(getString(R.string.action_view_preferences)).withSelectable(false).withIcon(GoogleMaterial.Icon.gmd_settings));
+        listItem.add(new SecondaryDrawerItem().withIdentifier(DRAWER_ITEM_ABOUT).withName(getString(R.string.about_title_activity)).withSelectable(false).withIcon(GoogleMaterial.Icon.gmd_info));
+        listItem.add(new SecondaryDrawerItem().withIdentifier(DRAWER_ITEM_LOG_OUT).withName(getString(R.string.action_logout)).withSelectable(false).withIcon(GoogleMaterial.Icon.gmd_exit_to_app));
+
+        IDrawerItem[] array = new IDrawerItem[listItem.size()];
+        listItem.toArray(array); // fill the array
+
+        drawer = new DrawerBuilder()
+                .withActivity(this)
+                //.withToolbar(toolbar)
+                .withAccountHeader(headerResult)
+                .withHasStableIds(true)
+                .withSelectedItem(-1)
+                .addDrawerItems(array)
+                .withOnDrawerItemClickListener((view, position, drawerItem) -> {
+                    if (drawerItem != null) {
+                        long drawerItemIdentifier = drawerItem.getIdentifier();
+
+                        if (drawerItemIdentifier == DRAWER_ITEM_EDIT_PROFILE) {
+                            Intent intent = new Intent(MainActivity.this, EditProfileActivity.class);
+                            startActivity(intent);
+                        } else if (drawerItemIdentifier == DRAWER_ITEM_FAVOURITES) {
+                            Intent intent = new Intent(MainActivity.this, FavouritesActivity.class);
+                            startActivity(intent);
+                        } else if (drawerItemIdentifier == DRAWER_ITEM_MUTED_USERS) {
+                            Intent intent = new Intent(MainActivity.this, AccountListActivity.class);
+                            intent.putExtra("type", AccountListActivity.Type.MUTES);
+                            startActivity(intent);
+                        } else if (drawerItemIdentifier == DRAWER_ITEM_BLOCKED_USERS) {
+                            Intent intent = new Intent(MainActivity.this, AccountListActivity.class);
+                            intent.putExtra("type", AccountListActivity.Type.BLOCKS);
+                            startActivity(intent);
+                        } else if (drawerItemIdentifier == DRAWER_ITEM_SEARCH) {
+                            Intent intent = new Intent(MainActivity.this, SearchActivity.class);
+                            startActivity(intent);
+                        } else if (drawerItemIdentifier == DRAWER_ITEM_PREFERENCES) {
+                            Intent intent = new Intent(MainActivity.this, PreferencesActivity.class);
+                            startActivity(intent);
+                        } else if (drawerItemIdentifier == DRAWER_ITEM_ABOUT) {
+                            Intent intent = new Intent(MainActivity.this, AboutActivity.class);
+                            startActivity(intent);
+                        } else if (drawerItemIdentifier == DRAWER_ITEM_LOG_OUT) {
+                            logout();
+                        } else if (drawerItemIdentifier == DRAWER_ITEM_FOLLOW_REQUESTS) {
+                            Intent intent = new Intent(MainActivity.this, AccountListActivity.class);
+                            intent.putExtra("type", AccountListActivity.Type.FOLLOW_REQUESTS);
+                            startActivity(intent);
+                        } else if (drawerItemIdentifier == DRAWER_ITEM_SAVED_TOOT) {
+                            Intent intent = new Intent(MainActivity.this, SavedTootActivity.class);
+                            startActivity(intent);
+                        } else if (drawerItemIdentifier == DRAWER_ITEM_LISTS) {
+                            startActivity(ListsActivity.newIntent(this));
+                        }
+
+                    }
+
+                    return false;
+                })
+                .build();
+
+        if(BuildConfig.DEBUG) {
+            IDrawerItem debugItem = new SecondaryDrawerItem()
+                    .withIdentifier(1337)
+                    .withName("debug")
+                    .withDisabledTextColor(Color.GREEN)
+                    .withSelectable(false)
+                    .withEnabled(false);
+            drawer.addItem(debugItem);
+        }
+
+        updateProfiles();
+    }
+
+    private boolean handleProfileClick(IProfile profile, boolean current) {
+        AccountEntity activeAccount = TuskyApplication.getAccountManager().getActiveAccount();
+
+        //open profile when active image was clicked
+        if (current && activeAccount != null) {
+            Intent intent = new Intent(MainActivity.this, AccountActivity.class);
+            intent.putExtra("id", activeAccount.getAccountId());
+            startActivity(intent);
+            return true;
+        }
+        //open LoginActivity to add new account
+        if(profile.getIdentifier() == DRAWER_ITEM_ADD_ACCOUNT ) {
+            startActivity(LoginActivity.getIntent(this, true));
+            return true;
+        }
+        //change Account
+        changeAccount(profile.getIdentifier());
+        return false;
+    }
+
+
+    private void changeAccount(long newSelectedId) {
+        TuskyApplication.getAccountManager().setActiveAccount(newSelectedId);
+
+        Intent intent = new Intent(this, MainActivity.class);
+        intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TASK);
+        startActivity(intent);
+        finish();
+
+        overridePendingTransition(R.anim.explode, R.anim.explode);
+    }
+
+    private void logout() {
+
+        AccountEntity activeAccount = TuskyApplication.getAccountManager().getActiveAccount();
+
+        if(activeAccount != null) {
+
+            new AlertDialog.Builder(this)
+                    .setTitle(R.string.action_logout)
+                    .setMessage(getString(R.string.action_logout_confirm, activeAccount.getFullName()))
+                    .setPositiveButton(android.R.string.yes, (dialog, which) -> {
+
+                        AccountManager accountManager = TuskyApplication.getAccountManager();
+
+                        NotificationHelper.deleteNotificationChannelsForAccount(accountManager.getActiveAccount(), MainActivity.this);
+
+                        AccountEntity newAccount = accountManager.logActiveAccountOut();
+
+                        if (!NotificationHelper.areNotificationsEnabled(MainActivity.this)) disablePushNotifications();
+
+                        Intent intent;
+                        if (newAccount == null) {
+                            intent = LoginActivity.getIntent(MainActivity.this, false);
+                        } else {
+                            intent = new Intent(MainActivity.this, MainActivity.class);
+                        }
+                        startActivity(intent);
+                        finish();
+                    })
+                    .setNegativeButton(android.R.string.no, null)
+                    .show();
+        }
+    }
+
+    private void fetchUserInfo() {
+
+        mastodonApi.accountVerifyCredentials().enqueue(new Callback<Account>() {
+            @Override
+            public void onResponse(@NonNull Call<Account> call, @NonNull Response<Account> response) {
+                if (response.isSuccessful()) {
+                    onFetchUserInfoSuccess(response.body());
+                } else {
+                    onFetchUserInfoFailure(new Exception(response.message()));
+                }
+            }
+
+            @Override
+            public void onFailure(@NonNull Call<Account> call, @NonNull Throwable t) {
+                onFetchUserInfoFailure((Exception) t);
+            }
+        });
+    }
+
+    private void onFetchUserInfoSuccess(Account me) {
+        // Add the header image and avatar from the account, into the navigation drawer header.
+
+        ImageView background = headerResult.getHeaderBackgroundView();
+
+        Picasso.with(MainActivity.this)
+                .load(me.header)
+                .placeholder(R.drawable.account_header_default)
+                .into(background);
+
+        AccountManager am = TuskyApplication.getAccountManager();
+
+        am.updateActiveAccount(me);
+
+        NotificationHelper.createNotificationChannelsForAccount(am.getActiveAccount(), this);
+
+        // Show follow requests in the menu, if this is a locked account.
+        if (me.locked && drawer.getDrawerItem(DRAWER_ITEM_FOLLOW_REQUESTS) == null) {
+            PrimaryDrawerItem followRequestsItem = new PrimaryDrawerItem()
+                    .withIdentifier(DRAWER_ITEM_FOLLOW_REQUESTS)
+                    .withName(R.string.action_view_follow_requests)
+                    .withSelectable(false)
+                    .withIcon(GoogleMaterial.Icon.gmd_person_add);
+            drawer.addItemAtPosition(followRequestsItem, 3);
+        }
+
+        updateProfiles();
+
+    }
+
+    private void updateProfiles() {
+        AccountManager am = TuskyApplication.getAccountManager();
+
+        List<AccountEntity> allAccounts = am.getAllAccountsOrderedByActive();
+
+        //remove profiles before adding them again to avoid duplicates
+        List<IProfile> profiles = new ArrayList<>(headerResult.getProfiles());
+        for(IProfile profile: profiles) {
+            if(profile.getIdentifier() != DRAWER_ITEM_ADD_ACCOUNT) {
+                headerResult.removeProfile(profile);
+            }
+        }
+
+        for(AccountEntity acc: allAccounts) {
+            headerResult.addProfiles(
+                    new ProfileDrawerItem()
+                            .withName(acc.getDisplayName())
+                            .withIcon(acc.getProfilePictureUrl())
+                            .withNameShown(true)
+                            .withIdentifier(acc.getId())
+                            .withEmail(acc.getFullName()));
+        }
+
+    }
+
+    private void onFetchUserInfoFailure(Exception exception) {
+        Log.e(TAG, "Failed to fetch user info. " + exception.getMessage());
+    }
+
+    @Nullable
+    @Override
+    public FloatingActionButton getActionButton() {
+        return composeButton;
     }
 }
