@@ -15,11 +15,12 @@
 
 package com.keylesspalace.tusky;
 
-import android.content.Context;
 import android.content.Intent;
-import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
+
+import com.keylesspalace.tusky.db.AccountEntity;
+import com.keylesspalace.tusky.util.NotificationHelper;
 
 public class SplashActivity extends AppCompatActivity {
     @Override
@@ -28,16 +29,16 @@ public class SplashActivity extends AppCompatActivity {
 
         /* Determine whether the user is currently logged in, and if so go ahead and load the
          * timeline. Otherwise, start the activity_login screen. */
-        SharedPreferences preferences = getSharedPreferences(
-                getString(R.string.preferences_file_key), Context.MODE_PRIVATE);
-        String domain = preferences.getString("domain", null);
-        String accessToken = preferences.getString("accessToken", null);
+
+        NotificationHelper.deleteLegacyNotificationChannels(this);
+
+        AccountEntity activeAccount = TuskyApplication.getAccountManager().getActiveAccount();
 
         Intent intent;
-        if (domain != null && accessToken != null) {
+        if (activeAccount != null) {
             intent = new Intent(this, MainActivity.class);
         } else {
-            intent = new Intent(this, LoginActivity.class);
+            intent = LoginActivity.getIntent(this, false);
         }
         startActivity(intent);
         finish();
