@@ -16,25 +16,44 @@
 package com.keylesspalace.tusky;
 
 import android.app.Application;
-import android.net.Uri;
+import android.app.UiModeManager;
+import android.arch.persistence.room.Room;
+import android.content.Context;
+import android.content.SharedPreferences;
+import android.preference.PreferenceManager;
+import android.support.v7.app.AppCompatDelegate;
 
-import com.squareup.picasso.Picasso;
+import com.evernote.android.job.JobManager;
 import com.jakewharton.picasso.OkHttp3Downloader;
+import com.keylesspalace.tusky.db.AccountManager;
+import com.keylesspalace.tusky.db.AppDatabase;
+import com.keylesspalace.tusky.util.OkHttpUtils;
+import com.keylesspalace.tusky.util.ThemeUtils;
+import com.squareup.picasso.Picasso;
 
 public class TuskyApplication extends Application {
+    public static final String APP_THEME_DEFAULT = ThemeUtils.THEME_NIGHT;
+
+    private static AppDatabase db;
+    private static AccountManager accountManager;
+
+    public static AppDatabase getDB() {
+        return db;
+    }
+
+    private static UiModeManager uiModeManager;
+
+    public static UiModeManager getUiModeManager() { return uiModeManager; }
+
     @Override
     public void onCreate() {
         super.onCreate();
         // Initialize Picasso configuration
         Picasso.Builder builder = new Picasso.Builder(this);
-        builder.downloader(new OkHttp3Downloader(this));
+        SharedPreferences preferences = PreferenceManager.getDefaultSharedPreferences(this);
+        builder.downloader(new OkHttp3Downloader(OkHttpUtils.getCompatibleClient(preferences)));
         if (BuildConfig.DEBUG) {
-            builder.listener(new Picasso.Listener() {
-                @Override
-                public void onImageLoadFailed(Picasso picasso, Uri uri, Exception exception) {
-                    exception.printStackTrace();
-                }
-            });
+            builder.listener((picasso, uri, exception) -> exception.printStackTrace());
         }
 
         try {
@@ -43,8 +62,23 @@ public class TuskyApplication extends Application {
             throw new RuntimeException(e);
         }
 
-        if (BuildConfig.DEBUG) {
-            Picasso.with(this).setLoggingEnabled(true);
-        }
+        db = Room.databaseBuilder(getApplicationContext(), AppDatabase.class, "tuskyDB")
+                .allowMainThreadQueries()
+                .addMigrations(AppDatabase.MIGRATION_2_3, AppDatabase.MIGRATION_3_4, AppDatabase.MIGRATION_4_5)
+                .build();
+
+        JobManager.create(this).addJobCreator(new NotificationPullJobCreator(this));
+
+        uiModeManager = (UiModeManager)getSystemService(Context.UI_MODE_SERVICE);
+
+        //necessary for Android < APi 21
+        AppCompatDelegate.setCompatVectorFromResourcesEnabled(true);
+
+        accountManager = new AccountManager();
     }
-}
+
+    public static AccountManager getAccountManager() {
+        return accountManager;
+    }
+
+ }
