@@ -20,6 +20,7 @@ import android.support.v7.widget.RecyclerView;
 
 import com.keylesspalace.tusky.entity.Account;
 import com.keylesspalace.tusky.interfaces.AccountActionListener;
+import com.keylesspalace.tusky.util.ListUtils;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -27,11 +28,16 @@ import java.util.List;
 public abstract class AccountAdapter extends RecyclerView.Adapter {
     List<Account> accountList;
     AccountActionListener accountActionListener;
+    FooterViewHolder.State footerState;
+
+    private String topId;
+    private String bottomId;
 
     AccountAdapter(AccountActionListener accountActionListener) {
         super();
         accountList = new ArrayList<>();
         this.accountActionListener = accountActionListener;
+        footerState = FooterViewHolder.State.END;
     }
 
     @Override
@@ -39,12 +45,17 @@ public abstract class AccountAdapter extends RecyclerView.Adapter {
         return accountList.size() + 1;
     }
 
-    public void update(List<Account> newAccounts) {
+    public void update(@Nullable List<Account> newAccounts, @Nullable String fromId,
+                       @Nullable String uptoId) {
         if (newAccounts == null || newAccounts.isEmpty()) {
             return;
         }
+
+        bottomId = fromId;
+        topId = uptoId;
+
         if (accountList.isEmpty()) {
-            accountList = newAccounts;
+            accountList = ListUtils.removeDuplicates(newAccounts);
         } else {
             int index = accountList.indexOf(newAccounts.get(newAccounts.size() - 1));
             for (int i = 0; i < index; i++) {
@@ -60,10 +71,25 @@ public abstract class AccountAdapter extends RecyclerView.Adapter {
         notifyDataSetChanged();
     }
 
-    public void addItems(List<Account> newAccounts) {
+    public void addItems(List<Account> newAccounts, @Nullable String fromId) {
+        if (fromId != null) {
+            bottomId = fromId;
+        }
         int end = accountList.size();
-        accountList.addAll(newAccounts);
-        notifyItemRangeInserted(end, newAccounts.size());
+        Account last = accountList.get(end - 1);
+        if (last != null && !findAccount(newAccounts, last.id)) {
+            accountList.addAll(newAccounts);
+            notifyItemRangeInserted(end, newAccounts.size());
+        }
+    }
+
+    private static boolean findAccount(List<Account> accounts, String id) {
+        for (Account account : accounts) {
+            if (account.id.equals(id)) {
+                return true;
+            }
+        }
+        return false;
     }
 
     @Nullable
@@ -84,10 +110,25 @@ public abstract class AccountAdapter extends RecyclerView.Adapter {
         notifyItemInserted(position);
     }
 
+    @Nullable
     public Account getItem(int position) {
         if (position >= 0 && position < accountList.size()) {
             return accountList.get(position);
         }
         return null;
+    }
+
+    public void setFooterState(FooterViewHolder.State newFooterState) {
+        footerState = newFooterState;
+    }
+
+    @Nullable
+    public String getBottomId() {
+        return bottomId;
+    }
+
+    @Nullable
+    public String getTopId() {
+        return topId;
     }
 }
