@@ -16,14 +16,13 @@
 package com.keylesspalace.tusky.entity;
 
 import android.os.Parcel;
+import android.os.Parcelable;
 import android.text.Spanned;
 
-import com.arlib.floatingsearchview.suggestions.model.SearchSuggestion;
 import com.google.gson.annotations.SerializedName;
 import com.keylesspalace.tusky.util.HtmlUtils;
-import com.keylesspalace.tusky.json.StringWithEmoji;
 
-public class Account implements SearchSuggestion {
+public class Account implements Parcelable {
     public String id;
 
     @SerializedName("username")
@@ -33,7 +32,7 @@ public class Account implements SearchSuggestion {
     public String username;
 
     @SerializedName("display_name")
-    public StringWithEmoji displayName;
+    public String displayName;
 
     public Spanned note;
 
@@ -71,15 +70,10 @@ public class Account implements SearchSuggestion {
     }
 
     public String getDisplayName() {
-        if (displayName.value.length() == 0) {
+        if (displayName.length() == 0) {
             return localUsername;
         }
-        return displayName.value;
-    }
-
-    @Override
-    public String getBody() {
-        return username;
+        return displayName;
     }
 
     @Override
@@ -92,7 +86,7 @@ public class Account implements SearchSuggestion {
         dest.writeString(id);
         dest.writeString(localUsername);
         dest.writeString(username);
-        dest.writeString(displayName.value);
+        dest.writeString(displayName);
         dest.writeString(HtmlUtils.toHtml(note));
         dest.writeString(url);
         dest.writeString(avatar);
@@ -103,15 +97,13 @@ public class Account implements SearchSuggestion {
         dest.writeString(statusesCount);
     }
 
-    public Account() {
-
-    }
+    public Account() {}
 
     protected Account(Parcel in) {
         id = in.readString();
         localUsername = in.readString();
         username = in.readString();
-        displayName = new StringWithEmoji(in.readString());
+        displayName = in.readString();
         note = HtmlUtils.fromHtml(in.readString());
         url = in.readString();
         avatar = in.readString();
