@@ -18,18 +18,17 @@ package com.keylesspalace.tusky.view;
 import android.content.Context;
 import android.support.v13.view.inputmethod.EditorInfoCompat;
 import android.support.v13.view.inputmethod.InputConnectionCompat;
-import android.support.v7.widget.AppCompatEditText;
+import android.support.v7.widget.AppCompatMultiAutoCompleteTextView;
 import android.util.AttributeSet;
 import android.view.inputmethod.EditorInfo;
 import android.view.inputmethod.InputConnection;
 
 import com.keylesspalace.tusky.util.Assert;
 
-public class EditTextTyped extends AppCompatEditText {
+public class EditTextTyped extends AppCompatMultiAutoCompleteTextView {
 
-    InputConnectionCompat.OnCommitContentListener onCommitContentListener;
-    String[] mimeTypes;
-    private OnPasteListener mOnPasteListener;
+    private InputConnectionCompat.OnCommitContentListener onCommitContentListener;
+    private String[] mimeTypes;
 
     public EditTextTyped(Context context) {
         super(context);
@@ -37,10 +36,6 @@ public class EditTextTyped extends AppCompatEditText {
 
     public EditTextTyped(Context context, AttributeSet attributeSet) {
         super(context, attributeSet);
-    }
-
-    public void addOnPasteListener(OnPasteListener mOnPasteListener) {
-        this.mOnPasteListener = mOnPasteListener;
     }
 
     public void setMimeTypes(String[] types,
@@ -60,27 +55,5 @@ public class EditTextTyped extends AppCompatEditText {
         } else {
             return connection;
         }
-    }
-
-    @Override
-    public boolean onTextContextMenuItem(int id) {
-        boolean consumed = super.onTextContextMenuItem(id);
-        switch (id) {
-            case android.R.id.paste:
-                onPaste();
-        }
-        return consumed;
-    }
-
-    /**
-     * Text was pasted into the EditText.
-     */
-    public void onPaste() {
-        if (mOnPasteListener != null)
-            mOnPasteListener.onPaste();
-    }
-
-    public interface OnPasteListener {
-        void onPaste();
     }
 }
